+++ apps/demo/src/pages/api/preview.ts
@@ -1,0 +1,14 @@
+import {backend} from '@alinea/content/backend.js'
+import type {NextApiRequest, NextApiResponse} from 'next'
+
+export default async function handler(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
+  const previewToken = decodeURIComponent(
+    new URL(req.url!, 'http://localhost').search
+  ).substring(1)
+  const {url} = await backend.parsePreviewToken(previewToken)
+  res.setPreviewData(previewToken)
+  res.redirect(url)
+}