+++ packages/dashboard/src/view/preview/BrowserPreview.tsx
@@ -1,3 +1,4 @@
+import {PreviewAction} from '@alinea/preview/RegisterPreview'
 import {
   AppBar,
   fromModule,
@@ -25,22 +26,26 @@ export type BrowserPreviewProps = {
   reload?: boolean
 }
 
-export function BrowserPreview({
-  url,
-  prettyUrl,
-  reload = true
-}: BrowserPreviewProps) {
+export function BrowserPreview({url, prettyUrl, reload}: BrowserPreviewProps) {
   const ref = useRef<HTMLIFrameElement>(null)
   const drafts = useDrafts()
   const status = useObservable(drafts.status)
   const [loading, setLoading] = useState(true)
-  const [hasHistory, setHasHistory] = useState(true)
+  const hasPreviewListener = useRef(false)
   useEffect(() => {
     setLoading(true)
+
+    function handleMessage(event: MessageEvent) {
+      if (event.data === PreviewAction.Pong) hasPreviewListener.current = true
+    }
+
+    addEventListener('message', handleMessage)
+    return () => removeEventListener('message', handleMessage)
   }, [url])
   useEffect(() => {
-    if (reload && status === DraftsStatus.Synced) {
-      ref.current?.setAttribute('src', url)
+    if (status === DraftsStatus.Synced) {
+      if (!reload && hasPreviewListener.current) handleRefetch()
+      else handleReload()
     }
   }, [status])
   useEffect(() => {
@@ -49,35 +54,48 @@ export function BrowserPreview({
       return () => clearTimeout(timeout)
     }
   }, [loading])
+
+  function post(action: PreviewAction) {
+    ref.current?.contentWindow?.postMessage(action, url)
+  }
+
+  function handlePrevious() {
+    post(PreviewAction.Previous)
+  }
+
+  function handleNext() {
+    post(PreviewAction.Next)
+  }
+
+  function handleReload() {
+    if (hasPreviewListener.current) post(PreviewAction.Reload)
+    else ref.current?.setAttribute('src', url)
+    setLoading(true)
+  }
+
+  function handleRefetch() {
+    post(PreviewAction.Refetch)
+  }
   return (
     <Preview>
       <div className={styles.root()}>
         <AppBar.Root>
-          {hasHistory && (
-            <>
-              <AppBar.Item
-                as="button"
-                icon={IcRoundArrowBack}
-                onClick={() => {
-                  ref.current?.contentWindow?.history.back()
-                }}
-              />
-              <AppBar.Item
-                as="button"
-                icon={IcRoundArrowForward}
-                onClick={() => {
-                  ref.current?.contentWindow?.history.forward()
-                }}
-              />
-            </>
-          )}
+          <>
+            <AppBar.Item
+              as="button"
+              icon={IcRoundArrowBack}
+              onClick={handlePrevious}
+            />
+            <AppBar.Item
+              as="button"
+              icon={IcRoundArrowForward}
+              onClick={handleNext}
+            />
+          </>
           <AppBar.Item
             as="button"
             icon={IcRoundRefresh}
-            onClick={() => {
-              if (hasHistory) ref.current?.contentWindow?.location.reload()
-              else ref.current?.setAttribute('src', url)
-            }}
+            onClick={handleReload}
           />
           <AppBar.Item full style={{flexGrow: 1, minWidth: 0}}>
             <Typo.Monospace
@@ -125,10 +143,11 @@ export function BrowserPreview({
             onLoad={() => {
               setLoading(false)
               try {
-                ref.current?.contentWindow?.history
-                setHasHistory(true)
+                const contentWindow = ref.current?.contentWindow
+                if (contentWindow)
+                  contentWindow.postMessage(PreviewAction.Ping, url)
               } catch (e) {
-                setHasHistory(false)
+                hasPreviewListener.current = false
               }
             }}
           />