+++ packages/ui/src/RichText.tsx
@@ -1,5 +1,5 @@
+import {ComponentType, Fragment, ReactElement, isValidElement} from 'react'
 import {HasType, TextDoc, TextNode, TypesOf} from '@alinea/core'
-import {ComponentType, Fragment, isValidElement, ReactElement} from 'react'
 
 export enum Elements {
   h1 = 'h1',
@@ -15,7 +15,9 @@ export enum Elements {
   ol = 'ol',
   li = 'li',
   a = 'a',
-  hr = 'hr'
+  hr = 'hr',
+  br = 'br',
+  small = 'small'
 }
 
 type Element = keyof typeof Elements
@@ -48,6 +50,10 @@ function nodeElement(
       return <blockquote style={style} />
     case 'horizontalRule':
       return <hr />
+    case 'hardBreak':
+      return <br />
+    case 'small':
+      return <small />
   }
 }