+++ packages/ui/src/UIStory.tsx
@@ -1,16 +1,19 @@
 import {PropsWithChildren} from 'react'
+import {PreferencesProvider} from './hook/UsePreferences'
 import {Viewport} from './Viewport'
 
 export function UIStory({children}: PropsWithChildren<{}>) {
   return (
-    <Viewport attachToBody>
-      <div
-        style={{
-          margin: 'auto'
-        }}
-      >
-        {children}
-      </div>
-    </Viewport>
+    <PreferencesProvider>
+      <Viewport attachToBody>
+        <div
+          style={{
+            margin: 'auto'
+          }}
+        >
+          {children}
+        </div>
+      </Viewport>
+    </PreferencesProvider>
   )
 }